move plugin installation to it's own test_install_plugins.py file
