+++ tests/test_app/test_init.py
@@ -8,7 +8,7 @@ from unittest import TestCase
 import pytest
 
 from cogent3 import app_help, available_apps, get_app, open_data_store
-from cogent3.app.composable import LOADER, WRITER, define_app, is_app
+from cogent3.app.composable import LOADER, WRITER, is_app
 from cogent3.util.table import Table
